Move mandarin type to shared dir
