+++ app/(language)/mandarin/quiz/utils/definition-utils.ts
@@ -1,4 +1,4 @@
-import { MandarinDefinition } from "../types/mandarin";
+import { MandarinDefinition } from "../../api/mandarin";
 
 export function selectRandomDefinition(definitions: MandarinDefinition[]) {
   return definitions[Math.floor(Math.random() * definitions.length)];
