+++ app/(language)/mandarin/quiz/utils/generate-fill-blank.ts
@@ -1,5 +1,5 @@
-import { MandarinFillBlankCardParams } from "../types/card";
-import { MandarinDefinition } from "../types/mandarin";
+import { MandarinFillBlankCardParams } from "../cards/card";
+import { MandarinDefinition } from "../../api/mandarin";
 import { selectRandomDefinition } from "./definition-utils";
 
 export function generateFillBlank(
