+++ app/(language)/mandarin/quiz/utils/generate-match-definition.ts
@@ -1,5 +1,5 @@
-import { MandarinMatchDefinitionParams } from "../types/card";
-import { MandarinDefinition } from "../types/mandarin";
+import { MandarinMatchDefinitionParams } from "../cards/card";
+import { MandarinDefinition } from "../../api/mandarin";
 import { selectRandomDefinition } from "./definition-utils";
 
 export function generateMatchDefinition(
